Split up shared_access_signature.ex into different modules based on type
